Creating mock workouts. Prisma types. Debug pages.

+++ src/actions/gymDataActions.ts
@@ -6,30 +6,64 @@ import { DateTime } from "luxon";
 import { createId } from "@paralleldrive/cuid2";
 
 import { revalidatePath } from "next/cache";
-
-const translateError = (error: Prisma.PrismaClientKnownRequestError) => {
-  let message = "";
-
-  switch (error.code) {
-    case "P2002":
-      message = `Duplicate ${
-        error.meta ? `of ${error.meta.modelName}` : ""
-      } found.`;
-      break;
-
-    default:
-      message = `Something went wrong.`;
-      break;
-  }
-
-  return message;
-};
+import { translateError } from "./prismaErrorHandler";
 
 type CreateCategoryProps = {
   name: string;
   id?: string;
 };
 
+// Workout handlers
+// =================
+export const handleCreateWorkout = async (date: DateTime) => {
+  const workout: Prisma.WorkoutCreateInput = {
+    date: date.setZone("Europe/Warsaw").toUTC().toJSDate(),
+  };
+};
+
+export const handleReturnWorkouts = async () => {
+  const workouts = await prisma.workout.findMany({});
+
+  return workouts;
+};
+
+export const handleReturnWorkoutByDate = async (inputDate: Date) => {
+  const dateInLocal = DateTime.fromJSDate(inputDate).setZone("Europe/Warsaw");
+
+  if (!dateInLocal.isValid) {
+    return new Error("Invalid date");
+  }
+
+  const dayStart = dateInLocal.startOf("day").toUTC();
+  const dayEnd = dateInLocal.endOf("day").toUTC();
+
+  const workout = await prisma.workout.findFirst({
+    where: {
+      date: {
+        gte: dayStart.toJSDate(),
+        lte: dayEnd.toJSDate(),
+      },
+    },
+    include: {
+      exerciseInstances: {
+        include: {
+          sets: true,
+          exerciseType: {
+            include: {
+              category: true,
+              loadingType: true,
+            },
+          },
+        },
+      },
+    },
+  });
+
+  return workout;
+};
+
+// Category handlers
+// =================
 export const handleCreateCategory = async ({
   name,
   id,
@@ -67,6 +101,23 @@ export const handleReturnCategories = async () => {
 };
 
 export const handleDeleteCategoryById = async (id: string) => {
+  // Check if the category is empty
+
+  const exercisesWithCategory = await prisma.exerciseType.findMany({
+    where: {
+      categoryId: id,
+    },
+  });
+
+  // TODO: Delete the related ExerciseTypes and ExerciseInstances upon confirmation
+  if (exercisesWithCategory) {
+    console.log("Category not empty. Aborting!");
+    return {
+      status: 400,
+      message: "Category not empty. Aborting!",
+    };
+  }
+
   try {
     const categories = await prisma.category.delete({
       where: {
@@ -86,5 +137,81 @@ export const handleDeleteCategoryById = async (id: string) => {
       };
     }
     throw e;
+  }
+};
+
+// Create one of everything
+// =========================
+export const handleCreateEverything = async () => {
+  try {
+    // Define Exercise Sets
+    const exerciseSets: Prisma.ExerciseSetCreateWithoutExerciseInstanceInput[] =
+      [
+        {
+          id: createId(),
+          load: "20x20",
+          order: 1,
+          wasCompleted: true,
+        },
+        {
+          id: createId(),
+          load: "20x20",
+          order: 2,
+          wasCompleted: false,
+        },
+      ];
+
+    // Define Exercise Instance with nested Exercise Sets and Exercise Type
+    const exerciseInstance: Prisma.ExerciseInstanceCreateWithoutWorkoutInput = {
+      id: createId(),
+      order: 1,
+      sets: {
+        create: exerciseSets,
+      },
+      exerciseType: {
+        create: {
+          id: createId(),
+          personalBest: "0",
+          category: {
+            create: {
+              id: createId(),
+              name: `TestCategory_${Math.floor(
+                Math.random() * 1000
+              ).toString()}`,
+            },
+          },
+          loadingType: {
+            create: {
+              id: createId(),
+              areBothSignificant: true,
+              firstUnit: "kg",
+              secondUnit: "reps",
+            },
+          },
+        },
+      },
+    };
+
+    // Define Workout with nested Exercise Instance
+    const workout: Prisma.WorkoutCreateInput = {
+      date: new Date(),
+      exerciseInstances: {
+        create: [exerciseInstance],
+      },
+    };
+
+    // Execute the nested create operation
+    const createdWorkout = await prisma.workout.create({
+      data: workout,
+    });
+
+    console.log(
+      "Successfully created instances of all models:",
+      createdWorkout
+    );
+  } catch (error) {
+    console.error("Error creating instances:", error);
+  } finally {
+    await prisma.$disconnect();
   }
 };
